refactor: move modules out of entrypoint model
