+++ src/app/entrypoint/model/index.ts
@@ -1,3 +1,3 @@
-export * from './modules';
+export * from '../modules';
 export * from './lib';
 export type * from './types';
