+++ src/app/entrypoint/modules/scena-preview/lib/useScenaPreview.svelte.ts
@@ -2,7 +2,7 @@ import { ScenaEvent } from '@/entities/event';
 import type { ScenaEventEmitter } from '@/entities/event';
 import { OverrideLayer } from '@/shared/enums';
 
-import type { UseScenaConfigReturns } from '../../../types';
+import type { UseScenaConfigReturns } from '../../../model/types';
 import type { UseScenaOverridesReturns } from '../../scena-overrides';
 import type { UseScenaPreviewReturns } from '../types';
