+++ src/app/entrypoint/modules/scena-responsive/index.ts
